+++ core/src/crypto/mod.rs
@@ -74,6 +74,11 @@ pub fn concat_data_and_nonce(data: &[u8], nonce: &[u8]) -> Vec<u8> {
 
 /// Creates a `SignatureData` object by signing the supplied certificate and nonce with a pkey
 pub fn create_signature_data(signing_key: &PrivateKey, security_policy: SecurityPolicy, contained_cert: &ByteString, nonce: &ByteString) -> Result<SignatureData, StatusCode> {
+
+    // TODO this function should be refactored to return an error if the contained cert or nonce is incorrect, not a blank signature. That
+    //  very much depends on reading the spec to see what should happen if its not possible to create a signature, e.g. because
+    //  policy is None.
+
     let (algorithm, signature) = if contained_cert.is_null() || nonce.is_null() {
         (UAString::null(), ByteString::null())
     } else {